requirements.txt
# Core RAG and Document Processing
llama-index>=0.9.48
llama-parse>=0.3.4
pypdf>=4.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.3
unstructured>=0.12.4
llama-index-vector-stores-qdrant

# Vector Store
chromadb>=0.4.22
qdrant-client>=1.7.3

# LLM and Agents
langchain>=0.1.9
langchain-community>=0.0.24
langgraph>=0.0.26
transformers>=4.38.2
accelerate>=0.27.2
bitsandbytes>=0.42.0

# Web Search and Tools
serpapi>=0.1.0
tavily-python>=0.3.1
requests>=2.31.0

# API and Server
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
langchain-openai>=0.0.8
openai>=1.12.0
tiktoken>=0.6.0
aiohttp>=3.9.3
tenacity>=8.2.3
python-magic>=0.4.27
PyPDF2>=3.0.1
markdown>=3.5.2
nltk>=3.8.1
jwt
passlib
pydantic[email]

# Caching and Real-time
redis>=5.0.1
aioredis>=2.0.1
fastapi-cache2>=0.2.1
websockets>=12.0
channels>=4.0.0
channels-redis>=4.2.0

# UI and Analytics
streamlit>=1.31.1
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.4
scikit-learn>=1.4.0
wordcloud>=1.9.3
matplotlib>=3.8.3
seaborn>=0.13.2
google-search-results


# System monitoring
psutil>=5.9.8

# Testing
pytest>=8.0.2
pytest-asyncio>=0.23.5
httpx>=0.26.0
pytest-cov>=4.1.0

# Development
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0 